silant/backend/requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
rpds-py==0.22.3
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
typing_extensions==4.12.2
uritemplate==4.1.1
